+++ src/main/java/com/example/demo/model/Response/BookingResponse.java
@@ -18,7 +18,7 @@ public class BookingResponse {
     private long bookingId;
     private LocalDate startDate;
     private String status;
-    private float price;
+    private double price;
     private String duration;
     private String email;
     private String phone;
